+++ src/main/java/org/solitaire/spider/SpiderBoard.java
@@ -295,14 +295,14 @@ public class SpiderBoard extends GameBoard<Card[]> {
     }
 
     @Override
-    public double score() {
+    public int score() {
         if (super.score() == 0) {
             super.score(calcBoardScore());
         }
         return super.score();
     }
 
-    private double calcBoardScore() {
+    private int calcBoardScore() {
         candidates(findCandidates());
         return candidates().size();
     }
